[FE] modify: post interface 분리 및 경로 변경
