+++ frontend/src/apis/Api.ts
@@ -5,9 +5,8 @@ import api from "./BaseUrl";
 import { DetailConfig } from "pages/setting/DetailSettingPage";
 import { ModifiedPageConfig } from "pages/PageSettingPage";
 import { UserInfoConfig } from "interfaces/auth.interface";
-import { PostListConfig } from "stores/postStore";
 import { MyBlogInfoConfig, MyInfoConfig } from "interfaces/setting.interface";
-import { TempPostConfig, TempPostDetailConfig } from "interfaces/post.interface";
+import { PostListConfig, TempPostConfig, TempPostDetailConfig } from "interfaces/post.interface";
 
 type ServerResponse = {
   message?: string; // 메시지
