+++ frontend/src/interfaces/post.interface.ts
@@ -1,3 +1,27 @@
+export interface PostCommonConfig {
+  title: string;
+  description: string;
+  category: string;
+}
+
+export interface PostListConfig extends PostCommonConfig {
+  tag: string[];
+  date: string;
+  directory: string;
+  imgUrl: string;
+}
+
+export interface PostDetailConfig extends PostCommonConfig {
+  content: string;
+  tagList: string[];
+  fileList?: any[];
+  files?: any[];
+  images?: {
+    name: string;
+    url: string;
+  }[];
+}
+
 export interface TempPostConfig {
   tempPostId: number;
   title: string;
