+++ frontend/src/stores/postStore.ts
@@ -1,29 +1,6 @@
+import { PostDetailConfig, PostListConfig } from "interfaces/post.interface";
 import { create } from "zustand";
 import { immer } from "zustand/middleware/immer";
-
-interface PostCommonConfig {
-  title: string;
-  description: string;
-  category: string;
-}
-
-export interface PostListConfig extends PostCommonConfig {
-  tag: string[];
-  date: string;
-  directory: string;
-  imgUrl: string;
-}
-
-export interface PostDetailConfig extends PostCommonConfig {
-  content: string;
-  tagList: string[];
-  fileList?: any[];
-  files?: any[];
-  images?: {
-    name: string;
-    url: string;
-  }[];
-}
 
 interface State {
   editPost: PostDetailConfig;
